refactored statuspage App to components App->GroupCheckContainer->GroupCheck
